Hopefully correct c part on all files

+++ mr-pr-mpi-base.cpp
@@ -56,12 +56,27 @@ int main(int narg, char** args){
     auto start = std::chrono::high_resolution_clock::now();
     std::vector<std::string > v = split(file,'.');
     std::string filename = v[0];
+    // std::cout<<filename<<std::endl;
     MPI_Init(&narg,&args);
     double convergence = 0.0001;
-    int me,nprocs;
+    int me,nprocs,idk;
     MPI_Comm_rank(MPI_COMM_WORLD,&me);
     MPI_Comm_size(MPI_COMM_WORLD,&nprocs);
-
+    if(me==0){
+        std::string line;
+        std::ifstream infile(file);
+        int maxx(0);
+    	if(infile.is_open()){
+    		while(getline(infile,line)){
+    			std::vector<std::string> splittedString = split(line,' ');
+    			int k = std::max(std::stoi(splittedString[1]) , std::stoi(splittedString[0]));
+                if(k>maxx)
+                    maxx = k;
+    		}
+            infile.close();
+    	}
+        idk= maxx + 1;
+    }
     if(narg<=1){
         std::cout<<"File arguement not given"<<std::endl;
         MPI_Abort(MPI_COMM_WORLD,1);
@@ -74,20 +89,31 @@ int main(int narg, char** args){
     MapReduce *newImportances   = new MapReduce(MPI_COMM_WORLD);
     MapReduce *difference       = new MapReduce(MPI_COMM_WORLD);
     MPI_Barrier(MPI_COMM_WORLD);
-    nodes -> map(narg-1,&args[1],0,1,0,getNodes,NULL);
+    MPI_Bcast(&idk,1,MPI_INT,0,MPI_COMM_WORLD);
+    int* IndividualExistence    = new int[idk];
+    int* CollectiveExistence    = new int[idk];
+    for(int i(0);i<idk;i++){
+        IndividualExistence[i] = 0;
+        CollectiveExistence[i] = 0;
+    }
+    nodes -> map(narg-1,&args[1],0,1,0,getNodes,(void*) IndividualExistence);
     links -> map(narg-1,&args[1],0,1,0,getLinks,NULL);
+    MPI_Reduce(IndividualExistence, CollectiveExistence, idk, MPI_INT, MPI_MAX,0,MPI_COMM_WORLD);
+    MPI_Bcast(CollectiveExistence, idk, MPI_INT, 0, MPI_COMM_WORLD);
     int outgoingNodes = links -> collate(NULL);
     int totalNodes = nodes -> collate(NULL);
-    double *finalIndividualImportances = new double[totalNodes];
-    double *finalCollectiveImportances = new double[totalNodes];
+    double *finalIndividualImportances = new double[idk];
+    double *finalCollectiveImportances = new double[idk];
     for(int i(0);i<totalNodes;i++)finalIndividualImportances[i] = 0.0;
     int danglingNodes = totalNodes - outgoingNodes;
     double danglingImportance = 0.85*((double)danglingNodes/(((double)totalNodes)*((double)totalNodes))) + (0.15/((double)totalNodes));
     nodes -> reduce(initializeImportance,(void*) &totalNodes);
     newImportances ->map(nodes, copyImportances, NULL);
     links -> reduce(initializeGraph,(void*) &totalNodes);
+    // links-> map(links,outputValues,links);
     danglingNode -> map(nodes, initializeDanglingImportance, &danglingImportance);
     links -> add(danglingNode);
+    // links -> map(links,outputValues,NULL);
     links -> collate(NULL);
     S_and_alpha curr;
     curr.totalImportance    = 0;
@@ -99,8 +125,11 @@ int main(int narg, char** args){
     double collectiveDanglingImportance = 1.0;
     double individualDifference         = 0.0;
     double collectiveDifference         = 1.0;
-    while(true){;
+    int numIter                         = 0;
+    while(true){
+        numIter++;
         links           -> reduce(calculateImportances,(void*) &curr);
+        // std::cout<<"i am here"<<std::endl;
         individualTotalImportance    = curr.totalImportance;
         individualDanglingImportance = curr.danglingImportance;
         MPI_Reduce(&individualTotalImportance,&collectiveTotalImportance,1,MPI_DOUBLE,MPI_SUM,0,MPI_COMM_WORLD);
@@ -118,12 +147,15 @@ int main(int narg, char** args){
         difference      -> add(prevImportances);
         difference      -> collate(NULL);
         difference      -> reduce(calculateDifference, (void*) &individualDifference);
+        // std::cout << "calculated difference" << '\n';
         MPI_Reduce(&individualDifference,&collectiveDifference,1,MPI_DOUBLE,MPI_SUM,0,MPI_COMM_WORLD);
         MPI_Bcast(&collectiveDifference, 1, MPI_DOUBLE, 0, MPI_COMM_WORLD);
-        if(me == 0) std::cout<<collectiveDifference<<std::endl;
-        if(collectiveDifference<convergence){
+        // if(me == 0) std::cout<<collectiveDifference<<std::endl;
+        if(collectiveDifference<convergence || numIter>100){
+            // std::cout<<"Here's the segmentation fault"<<std::endl;
             newImportances -> map(newImportances, printValues, (void*)finalIndividualImportances);
-            MPI_Reduce(finalIndividualImportances,finalCollectiveImportances,totalNodes,MPI_DOUBLE,MPI_SUM,0,MPI_COMM_WORLD);
+            MPI_Reduce(finalIndividualImportances,finalCollectiveImportances,idk,MPI_DOUBLE,MPI_SUM,0,MPI_COMM_WORLD);
+            // std::cout<<"No segmentation fault"<<std::endl;
             if(me == 0){
                 std::string textfile = "";
                 textfile.append(filename);
@@ -131,9 +163,11 @@ int main(int narg, char** args){
                 std::ofstream outputFile;
                 outputFile.open(textfile);
                 double sum = 0.0;
-                for(int i(0);i<totalNodes;i++){
-                    outputFile << i <<" = "<<finalCollectiveImportances[i]<<std::endl;
-                    sum+=finalCollectiveImportances[i];
+                for(int i(0);i<idk;i++){
+                    if(CollectiveExistence[i]!=0){
+                        outputFile << i <<" = "<<finalCollectiveImportances[i]<<std::endl;
+                        sum+=finalCollectiveImportances[i];
+                    }
                 }
                 outputFile<<"sum = "<<sum<<std::endl;
             }
@@ -154,14 +188,17 @@ int main(int narg, char** args){
         collectiveTotalImportance   = 0.0;
         collectiveDanglingImportance= 0.0;
     }
+    if(me == 0){
+        auto stop = std::chrono::high_resolution_clock::now();
+        std::chrono::duration<double,std::milli> duration = (stop - start);
+        std::cout << duration.count()/1000.0<<" seconds taken." << '\n';
+    }
     MPI_Finalize();
-    auto stop = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double,std::milli> duration = (stop - start);
-    std::cout << duration.count()/1000.0<<" seconds taken." << '\n';
 }
 
 
 void getNodes(int itask, char* fname, KeyValue *kv, void *ptr){
+    int* doesExist = (int*) ptr;
     struct stat stbuf;
     int flag = stat(fname,&stbuf);
     if (flag < 0) {
@@ -181,6 +218,7 @@ void getNodes(int itask, char* fname, KeyValue *kv, void *ptr){
     while (word) {
         std::string str = word;
         int k = std::stoi(str) + 2;
+        doesExist[k-2] = 1;
         std::string f_str = std::to_string(k);
         char char_array[f_str.length() + 1];
         strcpy(char_array,f_str.c_str());
@@ -213,9 +251,20 @@ void getLinks(int itask, char* fname, KeyValue *kv, void *ptr){
         char char_array[f_str.length() + 1];
         strcpy(char_array,f_str.c_str());
         if(isReceivingNode){
+            std::string outlink = char_array;
+            std::string add = "00000000";
+            outlink = add.append(outlink);
+            if(outlink.length()>9){
+                int extraZeros = outlink.length()-9;
+                outlink = outlink.substr(extraZeros,outlink.length()-1);
+            }
+            char char_array[outlink.length()+1];
+            strcpy(char_array,outlink.c_str());
+            // std::cout<<outlink<<"\t"<<outlink.length()<<std::endl;
             char chararray[inNode.length() + 1];
             strcpy(chararray,inNode.c_str());
             kv->add(chararray,strlen(chararray)+1,char_array,strlen(char_array)+1);
+            // std::cout<<char_array<<" "<<strlen(char_array)+1<<std::endl;
             isReceivingNode = false;
         }
         else{
@@ -260,12 +309,18 @@ void initializeGraph(char *key, int keybytes, char *multivalue, int nvalues, int
     char chararray[str.length() + 1];
     strcpy(chararray,str.c_str());
     for(int i(0);i<nvalues;i++){
-        char currValue[nvalues];
+        char currValue[*valuebytes];
         for(int j(0);j< *valuebytes;j++){
             currValue[j] = multivalue[j+i*(*valuebytes)];
         }
-        kv->add(currValue,strlen(currValue)+1,char_array,16);
-        kv->add(currValue,strlen(currValue)+1,chararray,16);
+        std::string bad = currValue;
+        int helper = std::stoi(currValue);
+        std::string good = std::to_string(helper);
+        char currValue_good[good.length()+1];
+        strcpy(currValue_good, good.c_str());
+        kv->add(currValue_good,strlen(currValue_good)+1,char_array,16);
+        kv->add(currValue_good,strlen(currValue_good)+1,chararray,16);
+        // std::cout<<currValue<<std::endl;
     }
     double has_outLinks = 1.5;
     std::string f = std::to_string(has_outLinks);
@@ -274,6 +329,7 @@ void initializeGraph(char *key, int keybytes, char *multivalue, int nvalues, int
     char charArray[f.length() + 1];
     strcpy(charArray,f.c_str());
     kv->add(key,keybytes,charArray,strlen(charArray)+1);
+    // std::cout<<key<<std::endl;
 }
 void emitImportances(char *key, int keybytes, char *multivalue, int nvalues, int *valuebytes, KeyValue *kv, void *ptr){
     // std::cout<<"I reached here!!!!"<<std::endl;
@@ -365,7 +421,10 @@ void calculateDifference(char *key, int keybytes, char *multivalue, int nvalues,
 void calculateImportances(char *key, int keybytes, char *multivalue, int nvalues, int *valuebytes, KeyValue *kv, void *ptr){
     S_and_alpha* s_and_alpha = (S_and_alpha*) ptr;
     std::string current_key = key;
+    // std::cout<<"I am above"<<std::endl;
+    // std::cout<<current_key<<std::endl;
     double inNode = (double)std::stoi(current_key);
+    // std::cout<<"I am below"<<std::endl;
     std::string ff = std::to_string(inNode);
     std::string zeros = "0000000";
     ff = zeros.append(ff);
